requirements.txt
annotated-types==0.7.0
anyio==4.11.0
azure-common==1.1.28
azure-core==1.36.0
azure-cosmos==4.5.1
azure-search-documents==11.7.0b1
azure-storage-blob==12.19.0
beautifulsoup4==4.14.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
dotenv==0.9.9
et_xmlfile==2.0.0
fastapi==0.118.3
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.10
iniconfig==2.3.0
isodate==0.7.2
jiter==0.11.1
lxml==6.0.2
numpy==2.3.4
openai==1.97.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==12.0.0
pluggy==1.6.0
pycparser==2.23
pydantic==2.12.0
pydantic-settings==2.11.0
pydantic_core==2.41.1
Pygments==2.19.2
pypdfium2==5.0.0
pytest==8.4.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
num2words==0.5.13
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sse-starlette==2.0.0
starlette==0.48.0
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
watchfiles==1.1.1
websockets==15.0.1
wheel==0.45.1
xlsxwriter==3.2.9
jinja2==3.1.4
reportlab==4.2.5
markdown==3.7
playwright

# Web Search & Deep Research dependencies
firecrawl-py==4.6.0
duckduckgo-search==7.5.4
markdownify==1.1.0
litellm==1.75.0
json5==0.12.0
google-search-results==2.4.2

# Knowledge Base dependencies
sqlalchemy==2.0.44
PyPDF2==3.0.1